# optimization: avoid merging single profiles