There are code paths where we end up merging a single profile. While in most cases the overhead is low, this is not the case when we are dealing with hundreds of thousands of samples. 

Example:

```
var pm pprof.ProfileMerge
profile := xxx // can come from different sources
pm.merge(profile)
return p.Profile()
```

We end up doing such merges when
- resolving symbols, and there is a single partition
- aggregating query reports, and there is a single report

End-to-end benchmark, with individual profiles having ~100-500k samples, ~50-100k locations, ~10-20k unique strings:

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/pyroscope/pkg/querybackend
cpu: Apple M2 Pro
                                      │ reference.txt │           optimized.txt           │
                                      │    sec/op     │   sec/op    vs base               │
_Invoke_QueryPprof/with_raw_output-10     8.485 ± 14%   7.861 ± 4%  -7.35% (p=0.004 n=10)

                                      │ reference.txt │            optimized.txt            │
                                      │     B/op      │     B/op      vs base               │
_Invoke_QueryPprof/with_raw_output-10    14.84Gi ± 2%   13.48Gi ± 3%  -9.18% (p=0.000 n=10)

                                      │ reference.txt │            optimized.txt            │
                                      │   allocs/op   │  allocs/op   vs base                │
_Invoke_QueryPprof/with_raw_output-10     73.24M ± 0%   65.79M ± 0%  -10.17% (p=0.000 n=10)
```